Change energy_group type to double

+++ src/physics/degas2/DG2Physics.h
@@ -18,7 +18,7 @@ struct ParticleInfo {
   double position[3];  // Position in space (x, y, z)
   double direction[3]; // Direction vector (unit vector)
   double weight;
-  int energy_group; // Energy group *index*
+  double energy_group; // Energy group *index*
 };
 
 struct FieldInfo {
